chore: full types refactor
