+++ custom-configuration.d.ts
@@ -1,3 +1,3 @@
-import { ExternalVuetifyOptions } from './dist/types';
+import type { ExternalVuetifyOptions } from 'vuetify-nuxt-module';
 declare function defineVuetifyConfiguration(vuetifyOptions: ExternalVuetifyOptions): ExternalVuetifyOptions;
 export { defineVuetifyConfiguration };
