+++ src/runtime/plugins/vuetify-client-hints.server.ts
@@ -91,7 +91,7 @@ const plugin: Plugin<{
 
   return {
     provide: reactive({
-      ssrClientHints: state.value,
+      ssrClientHints: state,
     }),
   }
 })
