+++ src/types.ts
@@ -1,4 +1,5 @@
 import type { LocaleOptions, RtlOptions, VuetifyOptions } from 'vuetify'
+import type { HookResult } from '@nuxt/schema'
 
 export type DateAdapter = 'vuetify' | 'date-fns' | 'moment' | 'luxon' | 'dayjs' | 'js-joda' | 'date-fns-jalali' | 'jalaali' | 'hijri' | 'custom'
 
@@ -398,3 +399,37 @@ export interface SSRClientHintsConfiguration {
     lightThemeName: string
   }
 }
+
+export interface ModuleHooks {
+  'vuetify:registerModule': (registerModule: (config: InlineModuleOptions) => void) => HookResult
+}
+/*
+declare module '#app' {
+  interface RuntimeNuxtHooks {
+    'vuetify:configuration': (options: {
+      isDev: boolean
+      vuetifyOptions: VuetifyOptions
+    }) => HookResult
+    'vuetify:before-create': (options: {
+      isDev: boolean
+      vuetifyOptions: VuetifyOptions
+    }) => HookResult
+    'vuetify:ready': (vuetify: ReturnType<typeof createVuetify>) => HookResult
+    'vuetify:ssr-client-hints': (options: {
+      vuetifyOptions: VuetifyOptions
+      ssrClientHints: SSRClientHints
+      ssrClientHintsConfiguration: SSRClientHintsConfiguration
+    }) => HookResult
+  }
+}
+
+declare module '@nuxt/schema' {
+  interface NuxtConfig {
+    ['vuetify']?: Partial<ModuleOptions>
+  }
+  interface NuxtOptions {
+    ['vuetify']?: ModuleOptions
+  }
+  interface NuxtHooks extends ModuleHooks {}
+}
+*/
