+++ src/utils/configure-nuxt.ts
@@ -1,5 +1,5 @@
 import type { Nuxt } from '@nuxt/schema'
-import { addImports, addPlugin, extendWebpackConfig } from '@nuxt/kit'
+import { addImports, addPlugin, addTemplate, extendWebpackConfig } from '@nuxt/kit'
 import { transformAssetUrls } from 'vite-plugin-vuetify'
 import defu from 'defu'
 import type { VuetifyNuxtContext } from './config'
@@ -46,10 +46,58 @@ export function configureNuxt(
     throw new Error('Webpack is not supported: vuetify-nuxt-module module can only be used with Vite!')
   })
 
+  addTemplate({
+    filename: 'module/vuetify-nuxt-module.d.ts',
+    getContents: () => `// Generated by vuetify-nuxt-module
+import type {
+  InlineModuleOptions,
+  ModuleOptions,
+  ModuleHooks,
+  SSRClientHints,
+  SSRClientHintsConfiguration
+} from 'vuetify-nuxt-module'
+import type { createVuetify, VuetifyOptions } from 'vuetify'
+import type { HookResult } from '@nuxt/schema'
+
+declare module '#app' {
+    interface RuntimeNuxtHooks {
+      'vuetify:configuration': (options: {
+        isDev: boolean
+        vuetifyOptions: VuetifyOptions
+      }) => HookResult
+      'vuetify:before-create': (options: {
+        isDev: boolean
+        vuetifyOptions: VuetifyOptions
+      }) => HookResult
+      'vuetify:ready': (vuetify: ReturnType<typeof createVuetify>) => HookResult
+      'vuetify:ssr-client-hints': (options: {
+        vuetifyOptions: VuetifyOptions
+        ssrClientHints: SSRClientHints
+        ssrClientHintsConfiguration: SSRClientHintsConfiguration
+      }) => HookResult
+    }
+}
+declare module '@nuxt/schema' {
+  interface NuxtConfig { ['vuetify']?: Partial<ModuleOptions> }
+  interface NuxtOptions { ['vuetify']?: ModuleOptions }
+  interface NuxtHooks {
+    'vuetify:registerModule': (registerModule: (config: InlineModuleOptions) => void) => HookResult
+  }
+}
+
+export {}
+`,
+  })
+
   nuxt.hook('prepare:types', ({ references }) => {
     references.push({ types: 'vuetify' })
     references.push({ types: 'vuetify-nuxt-module/custom-configuration' })
     references.push({ types: 'vuetify-nuxt-module/configuration' })
+    // let configuration = ctx.resolver.resolve('../configuration.d.ts')
+    // references.push({ path: ctx.resolver.resolve(nuxt.options.buildDir, configuration) })
+    // configuration = ctx.resolver.resolve('../custom-configuration.d.ts')
+    // references.push({ path: ctx.resolver.resolve(nuxt.options.buildDir, configuration) })
+    references.push({ path: ctx.resolver.resolve(nuxt.options.buildDir, 'module/vuetify-nuxt-module.d.ts') })
   })
 
   /* nuxt.hook('components:extend', async (c) => {
